separate data managing to another file
