+++ kono-judge-status/src/components/page/LandingPage.js
@@ -1,9 +1,10 @@
 import React, { useState, useContext, useEffect } from 'react';
 import styles from 'styles/components/LandingPage.module.scss';
+import classnames from 'lib/classnames';
+import { TIME_FILTERS } from 'lib/DataManaging';
 import { WebsocketController } from 'components/websocket';
 import { FilterSelector, DeviceSelector, DataPanel } from 'components/data';
-import { TIME_FILTERS, DataContext } from 'components/provider/DataProvider';
-import classnames from 'lib/classnames';
+import { DataContext } from 'components/provider/DataProvider';
 
 export default () => {
